Refactored file structure.
